no decimals on pixel css

+++ src/blagario/AgarElements/AgarElement.cs
@@ -35,10 +35,10 @@ namespace blagario.elements
         public Universe Universe {get; protected set;}
         public string CssClass => this.GetType().Name.ToLower();
         public virtual string CssStyle( Player c) => $@"
-            top: {c.YGame2World(CssY).ToString()}px ;
-            left: {c.XGame2World(CssX).ToString()}px ;
-            width: {(Diameter * c.Zoom).ToString()}px ;
-            height: {(Diameter * c.Zoom).ToString()}px ;
+            top: {(c.YGame2World(CssY)).ToString()}px ;
+            left: {(c.XGame2World(CssX)).ToString()}px ;
+            width: {((long)(Diameter * c.Zoom)).ToString()}px ;
+            height: {((long)(Diameter * c.Zoom)).ToString()}px ;
             ";
     }
 }
